refactor for test focused development
