+++ app/lib/api.js
@@ -1,9 +1,8 @@
 
-const endpoint = 'https://api.pexels.com/v1/collections/ylxa5x6?';
-const endpoint2 = 'https://api.pexels.com/v1/search?'
+
 
 export async function fetchImages() {
-
+  const endpoint = 'https://api.pexels.com/v1/collections/ylxa5x6?';
   const headers =  {
     'Content-Type': 'application/json',
     'Authorization': process.env.KEY,
@@ -20,6 +19,7 @@ export async function fetchImages() {
   try {
     const res = await fetch(endpoint + searchParams.toString(), {headers});
     const data = await res.json();
+    console.log(data);
     return data;
     
   } catch (error) {
@@ -29,7 +29,7 @@ export async function fetchImages() {
 }
 
 export async function fetchCategoryImages(category) {
-  await new Promise((resolve) => setTimeout(resolve, 3000));
+  const endpoint2 = 'https://api.pexels.com/v1/search?'
   
   const headers =  {
     'Content-Type': 'application/json',
@@ -39,7 +39,7 @@ export async function fetchCategoryImages(category) {
   const params = {
     query: `${category}`,
     page: '1',
-    per_page: '24',
+    per_page: '12',
   }
 
   const searchParams = new URLSearchParams(params);
